+++ freppledb/common/tests/frepplePages/frepplepage.py
@@ -17,12 +17,11 @@
 
 from freppledb.common.tests.frepplePages.freppleelement import (
     BasePageElement,
-)  # in here, import the class containing all the elements from your target page
+)
 from freppledb.common.tests.frepplePages.frepplelocators import (
     TableLocators,
     BasePageLocators,
-)  # here, we should find all the locators for your target page
-
+)
 
 import time
 
@@ -33,6 +32,7 @@ from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.common.action_chains import ActionChains
 
 from django.conf import settings
+from django.utils.formats import date_format
 
 ### Special page for common actions only
 
@@ -156,16 +156,14 @@ class TablePage(BasePage):
         time.sleep(0.3)
 
     def enter_text_in_inputdatefield(self, targetinputdatefield, newdate):
+        val = date_format(newdate, "DATETIME_FORMAT", use_l10n=False)
         targetinputdatefield.clear()
         time.sleep(0.3)
-        targetinputdatefield.send_keys(newdate.strftime("%Y-%m-%d %H:%M:%S"))
+        targetinputdatefield.send_keys(val)
         time.sleep(0.3)
         targetinputdatefield.send_keys(Keys.RETURN)
         time.sleep(0.3)
-        return newdate.strftime("%Y-%m-%d %H:%M:%S")
-
-    # def is_title_matches(self): # change to current url
-    #    return "Purchase orders" in self.driver.title
+        return val
 
     def click_save_button(self):
         save_button = self.driver.find_element(*TableLocators.TABLE_SAVE_BUTTON)