+++ freppledb/input/views/manufacturing.py
@@ -796,7 +796,6 @@ class CalendarDetail(GridReport):
         # Build up event list
         events = []
         curDate = request.report_startdate
-        prevDate = request.report_startdate
         curBucket = reportclass.findBucket(curDate, buckets)
         curPriority = curBucket.priority if curBucket else maxsize
         lastPriority = curPriority
@@ -1865,12 +1864,14 @@ class ManufacturingOrderList(OperationPlanMixin):
         GridFieldDateTime(
             "startdate",
             title=_("start date"),
-            extra='"formatoptions":{"srcformat":"Y-m-d H:i:s","newformat":"Y-m-d H:i:s", "defaultValue":""}, "summaryType":"min"',
+            extra='"formatoptions":{"srcformat":"Y-m-d H:i:s","newformat":"%s", "defaultValue":""}, "summaryType":"min"'
+            % settings.DATETIME_FORMAT,
         ),
         GridFieldDateTime(
             "enddate",
             title=_("end date"),
-            extra='"formatoptions":{"srcformat":"Y-m-d H:i:s","newformat":"Y-m-d H:i:s", "defaultValue":""}, "summaryType":"max"',
+            extra='"formatoptions":{"srcformat":"Y-m-d H:i:s","newformat":"%s", "defaultValue":""}, "summaryType":"max"'
+            % settings.DATETIME_FORMAT,
         ),
         GridFieldDuration(
             "opplan_duration",